+++ CHANGELOG.md
@@ -74,3 +74,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [0.2.1]: https://github.com/FranklinZEN/ThinkStash_KB/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/FranklinZEN/ThinkStash_KB/releases/tag/v0.2.0 
+
+
+- Proceed a test for Google CI/CD